+++ lab-patrick/routes/pic-routes.js
@@ -1,69 +1,19 @@
 'use strict';
 
-const fs = require('fs');
-const del = require('del');
-const path = require('path');
-const AWS  = require('aws-sdk');
 const multer = require('multer');
 const dataDir = `${__dirname}/../data`;
 const upload = multer({dest: dataDir});
-const createError = require('http-errors');
+// const createError = require('http-errors');
+
 const bearerAuth = require('../lib/bearer-auth-middleware');
-
-const Pic = require('../models/pic');
-const Gallery = require('../models/gallery');
-
-AWS.config.setPromisesDependency(require('bluebird'));
-
-const s3 = new AWS.S3();
+const picCtrl = require('../controller/pic-controller');
 
 
-function s3UploadProm(params) {
-  return new Promise((resolve, reject) => {
-    if(!params) return reject(new Error('params required'));
-    s3.upload(params, (err, data) => {
-      resolve(data);
-    });
-  });
-}
-
-module.exports = function(router) {
-  router.post('/gallery/:id/pic', bearerAuth, upload.single('image'), (req, res) => {
-
-    if(!req.file) return createError(400, 'Resource required');
-    if(!req.file.path) return createError(500, 'File not saved');
-
-    let ext = path.extname(req.file.originalname);
-    let params = {
-      ACL: 'public-read',
-      Bucket: process.env.AWS_BUCKET,
-      Key: `${req.file.filename}${ext}`,
-      Body: fs.createReadStream(req.file.path),
-    };
-
-    return Gallery.findById(req.params.id)
-    .then(() => s3UploadProm(params))
-    .then(s3Data => {
-      del([`${dataDir}/*`]);
-      let picData = {
-        name: req.body.name,
-        desc: req.body.desc,
-        userID: req.user._id,
-        galleryID: req.params.id,
-        imageURI: s3Data.Location,
-        objectKey: s3Data.Key,
-      };
-      return new Pic(picData).save();
-    })
+module.exports = function(router){
+  router.post('/gallery/:id/pic', bearerAuth, upload.single('image'), (req, res) =>{
+    picCtrl.createItem(req)
     .then(pic => res.json(pic))
-    .catch(err => res.send(err));
+    .catch(err => res.status(err.status).send(err.message));
   });
-
-  router.delete('/gallery/:id/pic', bearerAuth, (req, res)=>{
-    let params ={
-      Bucket: process.env.AWS_BUCKET,
-    }
-  })
-
   return router;
 };
